Consolidate early return checks into common functions that are used across command handlers
